chore: prepare for PyPi publishing
